+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - AssetsOnly attribute
 
 ### Changed
+- Improve TagAttributeDrawer class
 - Improve source code documentation
 
 ## [1.1.1] - 2020-06-12